packages/pybackend/pyproject.toml
[project]
name = "made-pybackend"
version = "0.1.0"
description = "FastAPI backend for MADE"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.111.0",
  "python-frontmatter==1.0.0",
  "uvicorn==0.29.0",
  "ruff>=0.1.0",
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "httpx>=0.24.0",
]

[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
exclude = [
  "/.venv",
  "/__pycache__",
  "*.pyc",
  "/tests",
  "/README.md",
  "/pyproject.toml",
  "/uv.lock"
]

[tool.uv]
package = true
